io: Make error module public and re-export errors no-inline
In io, the error module is currently private, with all errors
re-exported inline in the crate root. Following the standard set in the
other crates, errors should instead be located in a submodule and
re-exported with #[doc(no_inline)].

Move ReadError to error module.
Make error module pub.
re-export all error types in crate root with no_inline.
